@@ -3,26 +3,28 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@material-ui/core": "^4.12.4",
-    "@material-ui/icons": "^4.11.3",
+    "@mui/icons-material": "^6.4.11",
+    "@mui/material": "^6.4.11",
+    "@zxing/browser": "^0.1.5",
+    "@zxing/library": "^0.21.3",
     "qrcode": "^1.5.0",
-    "ra-i18n-polyglot": "^3.19.11",
-    "ra-language-finnish": "^3.18.3",
-    "react": "^16.14.0",
-    "react-admin": "^5.7.1",
-    "react-dom": "^16.14.0",
-    "react-qr-reader": "^2.2.1",
-    "react-router": "^5.3.3",
-    "react-router-navigation-prompt": "^1.9.6",
-    "react-scripts": "^5.0.1",
+    "ra-i18n-polyglot": "^5.7.3",
+    "ra-language-finnish": "^5.7.0",
+    "react": "^18.3.1",
+    "react-admin": "^5.7.3",
+    "react-dom": "^18.3.1",
+    "react-router-dom": "^6.30.0",
     "styled-components": "^4.4.1",
     "uifx": "^2.0.7"
   },
+  "devDependencies": {
+    "@vitejs/plugin-react": "^4.4.0",
+    "vite": "^6.3.1"
+  },
   "scripts": {
-    "start": "react-scripts start",
-    "build": "react-scripts build",
-    "test": "react-scripts test",
-    "eject": "react-scripts eject"
+    "dev": "vite",
+    "build": "vite build",
+    "preview": "vite preview"
   },
   "eslintConfig": {
     "extends": "react-app"
