@@ -1,16 +1,16 @@
 import React, { useState, useEffect } from 'react';
 import { Notification } from 'react-admin';
-import NavigationPrompt from "react-router-navigation-prompt";
-import ConfirmNavigationModal from "../ConfirmNavigationModal";
-import QrReader from 'react-qr-reader';
-import QrCheckResultScreen from "./qrCheckResultScreen.js";
-import LoadingMessage from "../loadingMessage";
+import { useBlocker, useLocation, useParams } from 'react-router-dom';
+import ConfirmNavigationModal from '../ConfirmNavigationModal';
+import QrReader from './qrReader';
+import QrCheckResultScreen from './qrCheckResultScreen';
+import LoadingMessage from '../loadingMessage';
 import { useNotify } from 'react-admin';
 import styled from 'styled-components';
 import { httpClientWithResponse } from '../../httpClients';
 import api from '../../api';
-import CheckinBackground from './checkInBackground.js';
-import { successSound, errorSound } from "../../audio/audio.js"
+import CheckinBackground from './checkInBackground';
+import { successSound, errorSound } from '../../audio/audio';
 
 const Container = styled.div`
   height: 100%;
@@ -25,43 +25,45 @@ const QrReaderContainer = styled.div`
   margin-top: 7.4em;
   width: 32em;
   border: 55px solid #f9e51e;
-  -webkit-box-shadow: 2px 10px 60px -19px rgba(0,0,0,0.75);
-  -moz-box-shadow: 2px 10px 60px -19px rgba(0,0,0,0.75);
-  box-shadow: 2px 10px 60px -19px rgba(0,0,0,0.75);
-`
+  -webkit-box-shadow: 2px 10px 60px -19px rgba(0, 0, 0, 0.75);
+  -moz-box-shadow: 2px 10px 60px -19px rgba(0, 0, 0, 0.75);
+  box-shadow: 2px 10px 60px -19px rgba(0, 0, 0, 0.75);
+`;
 
-let youthClubName = "";
+let youthClubName = '';
 
-const CheckInView = (props) => {
+const CheckInView = () => {
+  const location = useLocation();
+  const { youthClubId } = useParams();
   const [showQRCode, setShowQRCode] = useState(true);
   const [showQrCheckNotification, setShowQrCheckNotification] = useState(false);
   const [loading, setLoading] = useState(false);
   const [checkInSuccess, setCheckInSuccess] = useState(null);
   const notify = useNotify();
-  
+
   // Timeout for automatic logout
   const timeoutDuration = 1800; // Set your desired timeout duration in seconds
   const [remainingTime, setRemainingTime] = useState(timeoutDuration); // Set your desired initial timeout in seconds
-  
-  const logout = () => {
-    sessionStorage.removeItem("initialCheckin")
 
-    if (process.env.REACT_APP_ADMIN_FRONTEND_URL) {
-      document.location.href = process.env.REACT_APP_ADMIN_FRONTEND_URL;
+  const logout = () => {
+    sessionStorage.removeItem('initialCheckin');
+
+    if (import.meta.env.VITE_ADMIN_FRONTEND_URL) {
+      document.location.href = import.meta.env.VITE_ADMIN_FRONTEND_URL;
     } else {
-      document.location.href = "/";
+      document.location.href = '/';
     }
-  }
+  };
 
   // For scanning multiple users set a timeout for automatic logout
   // For scanning single user clear the session token and reload the page after 3 minutes to fix
   // qr reader stopping scanning after a period of time
   useEffect(() => {
-    const continuousCheckIn = sessionStorage.getItem("continuousCheckIn");
-    if (continuousCheckIn == "false") {
-      localStorage.removeItem("admin-token")
-      const initialCheckIn = sessionStorage.getItem("initialCheckIn");
-      const path = props.location.pathname;
+    const continuousCheckIn = sessionStorage.getItem('continuousCheckIn');
+    if (continuousCheckIn == 'false') {
+      localStorage.removeItem('admin-token');
+      const initialCheckIn = sessionStorage.getItem('initialCheckIn');
+      const path = location.pathname;
       const m = path.match(/\d+/);
       const id = m !== null ? m.shift() : null;
       if (id !== initialCheckIn) {
@@ -72,7 +74,9 @@ const CheckInView = (props) => {
       return () => clearTimeout(timer);
     } else {
       // Set up the timeout
-      const timer = setTimeout(() => { logout() }, timeoutDuration * 1000);
+      const timer = setTimeout(() => {
+        logout();
+      }, timeoutDuration * 1000);
 
       // Update the remaining time every second
       const interval = setInterval(() => {
@@ -85,17 +89,17 @@ const CheckInView = (props) => {
         clearInterval(interval);
       };
     }
-  }, [])
+  }, []);
 
   useEffect(() => {
-    if(props.location.state !== undefined) {
-      localStorage.setItem('youthClubName', JSON.stringify(props.location.state.record.name));
-      youthClubName = props.location.state.record.name;
+    if (location.state != null) {
+      localStorage.setItem('youthClubName', JSON.stringify(location.state.record.name));
+      youthClubName = location.state.record.name;
     }
-    if(props.location.state === undefined) {
-      youthClubName = JSON.parse(localStorage.getItem("youthClubName"));
+    if (location.state === undefined) {
+      youthClubName = JSON.parse(localStorage.getItem('youthClubName'));
     }
-  }, [])
+  }, []);
 
   const tryToPlayAudio = (success) => {
     if (success) {
@@ -111,8 +115,8 @@ const CheckInView = (props) => {
 
   const handleCheckInReturn = (success) => {
     setLoading(false);
-    setShowQRCode(false)
-    setCheckInSuccess(success)
+    setShowQRCode(false);
+    setCheckInSuccess(success);
     setShowQrCheckNotification(true);
     tryToPlayAudio(success).catch(() => notify('Audion toistaminen epäonnistui. Tarkista selaimesi oikeudet.', 'warning'));
     setTimeout(() => {
@@ -132,19 +136,19 @@ const CheckInView = (props) => {
 
       const url = api.youthClub.checkIn;
       const body = JSON.stringify({
-        clubId: props.match.params.youthClubId,
-        juniorId: qrData
+        clubId: youthClubId,
+        juniorId: qrData,
       });
       const options = {
         method: 'POST',
-        body
+        body,
       };
       await httpClientWithResponse(url, options, true)
         .then(response => {
           if (response.statusCode < 200 || response.statusCode >= 300) {
             setLoading(false);
-            notify('Jokin meni pieleen! Kokeile uudestaan.', 'warning')
-            setShowQRCode(true)
+            notify('Jokin meni pieleen! Kokeile uudestaan.', 'warning');
+            setShowQRCode(true);
           } else {
             handleCheckInReturn(response.success);
           }
@@ -152,37 +156,27 @@ const CheckInView = (props) => {
     }
   };
 
-  const handleError = () => {
-    notify('Jokin meni pieleen! Kokeile uudestaan.', 'warning')
-  };
-
   // Calculate minutes and seconds from remaining time
   const minutes = String(Math.floor(remainingTime / 60)).padStart(2, '0');
   const seconds = String(remainingTime % 60).padStart(2, '0');
 
-  const continuousCheckIn = sessionStorage.getItem("continuousCheckIn");
-  const facingMode = sessionStorage.getItem("facingMode") || "user";
+  const continuousCheckIn = sessionStorage.getItem('continuousCheckIn');
+  const facingMode = sessionStorage.getItem('facingMode') || 'user';
+
+  const blocker = useBlocker(continuousCheckIn == 'false');
 
   return (
     <Container>
       <Notification />
       <CheckinBackground />
 
-      {continuousCheckIn == "false" && (
-        <NavigationPrompt
-        afterConfirm={logout}
-        disableNative={true}
-        when={true}
-      >
-        {({ onConfirm, onCancel }) => (
-          <ConfirmNavigationModal
-            onCancel={onCancel}
-            onConfirm={onConfirm}
-          />
-        )}
-      </NavigationPrompt>
+      {blocker.state === 'blocked' && (
+        <ConfirmNavigationModal
+          onCancel={() => blocker.reset()}
+          onConfirm={() => logout()}
+        />
       )}
-      {continuousCheckIn == "true" && (
+      {continuousCheckIn == 'true' && (
         <p>Kirjaudutaan ulos: {minutes}:{seconds}</p>
       )}
 
@@ -191,9 +185,8 @@ const CheckInView = (props) => {
           <QrReader
             delay={300}
             onScan={handleScan}
-            onError={handleError}
             facingMode={facingMode}
-            style={{ width: "100%", height: "100%" }}
+            style={{ width: '100%', height: '100%' }}
           />
         </QrReaderContainer>
       )}
@@ -203,7 +196,7 @@ const CheckInView = (props) => {
         <LoadingMessage message={'Odota hetki'} />
       )}
     </Container>
-  )
+  );
 };
 
 export default CheckInView;
