@@ -1,5 +1,4 @@
 import React, { useState, useEffect } from 'react';
-import { useForm } from 'react-final-form';
 import QRCode from 'qrcode';
 import {
   List,
@@ -14,28 +13,32 @@ import {
   DateInput,
   BooleanInput,
   required,
+  Button,
   EditButton,
   Edit,
   Filter,
   useNotify,
   Pagination,
   FormDataConsumer,
+  useRecordContext,
 } from 'react-admin';
+import { useFormContext, useWatch } from 'react-hook-form';
 import {
   getYouthClubs,
   ageValidator,
   genderChoices,
   statusChoices,
 } from '../utils';
-import Button from '@material-ui/core/Button';
 import { httpClientWithResponse } from '../httpClients';
 import api from '../api';
 import usePermissions from '../hooks/usePermissions';
 import { hiddenFormFields } from '../customizations';
 
-const JuniorEditTitle = ({ record }) => (
-  <span>{`Muokkaa ${record.firstName} ${record.lastName}`}</span>
-);
+const JuniorEditTitle = () => {
+  const record = useRecordContext();
+  const name = record ? `${record.firstName} ${record.lastName}` : '';
+  return <span>Muokkaa {name}</span>;
+};
 
 const SMSwarning = () => (
   <div style={{ paddingTop: '1em', color: 'red' }}>
@@ -48,7 +51,6 @@ export const JuniorList = (props) => {
   const CustomPagination = (props) => (
     <Pagination rowsPerPageOptions={[5, 10, 25, 50]} {...props} />
   );
-  const notify = useNotify();
 
   const [youthClubs, setYouthClubs] = useState([]);
   useEffect(() => {
@@ -59,93 +61,12 @@ export const JuniorList = (props) => {
     addYouthClubsToState();
   }, []);
 
-  const JuniorFilter = (props) => (
-    <Filter {...props}>
-      <TextInput label="Nimi" source="name" autoFocus />
-      <TextInput label="Nuoren puhelinnumero" source="phoneNumber" autoFocus />
-      <TextInput
-        label="Huoltajan puhelinnumero"
-        source="parentsPhoneNumber"
-        autoFocus
-      />
-      <SelectInput
-        label="Kotinuorisotila"
-        source="homeYouthClub"
-        choices={youthClubs}
-      />
-      <SelectInput label="Tila" source="status" choices={statusChoices} />
-    </Filter>
-  );
-
-  const ResendSMSButton = (data) =>
-    data.record.status === 'accepted' || data.record.status === 'expired' ? (
-      <Button
-        size="small"
-        variant="contained"
-        onClick={() => resendSMS(data.record.phoneNumber)}
-      >
-        Lähetä SMS uudestaan
-      </Button>
-    ) : (
-      <Button disabled>Kotisoitto tekemättä</Button>
-    );
-
-  const generateQRAndOpen = async (id, owner) => {
-    try {
-      const data = await QRCode.toDataURL(id);
-      const image = new Image();
-      image.src = data;
-      image.width = 400;
-
-      const w = window.open('');
-      setTimeout(() => (w.document.title = `QR-koodi ${owner}`), 0);
-      w.document.write(image.outerHTML);
-      w.document.location = '#';
-      w.document.close();
-    } catch (err) {
-      alert('Virhe QR-koodin luonnissa');
-    }
-  };
-
-  const PrintQrCodeButton = (data) => (
-    <Button
-      size="small"
-      variant="contained"
-      onClick={() =>
-        generateQRAndOpen(
-          data.record.id,
-          `${data.record.firstName} ${data.record.lastName}`,
-        )
-      }
-    >
-      🔍QR
-    </Button>
-  );
-
-  const resendSMS = async (phoneNumber) => {
-    const url = api.junior.reset;
-    const body = JSON.stringify({
-      phoneNumber,
-    });
-    const options = {
-      method: 'POST',
-      body,
-    };
-    await httpClientWithResponse(url, options).then((response) => {
-      if (response.statusCode < 200 || response.statusCode >= 300) {
-        notify(response.message, 'warning');
-      } else {
-        notify(response.message);
-      }
-    });
-  };
-
   return (
     <List
       title="Nuoret"
       pagination={<CustomPagination />}
       debounce={1000}
-      filters={<JuniorFilter />}
+      filters={<JuniorFilter youthClubs={youthClubs} />}
       bulkActionButtons={false}
       exporter={false}
       {...props}
@@ -172,7 +93,7 @@ export const JuniorList = (props) => {
         />
         <DateField label="Päiväys" source="creationDate" locales={['fi']} />
         <SelectField label="Tila" source="status" choices={statusChoices} />
-        <PrintQrCodeButton></PrintQrCodeButton>
+        <PrintQrCodeButton />
         <ResendSMSButton />
         <EditButton className="focusable" />
       </Datagrid>
@@ -180,25 +101,111 @@ export const JuniorList = (props) => {
   );
 };
 
-const getDummyPhoneNumber = async (cb) => {
-  const url = api.junior.dummynumber;
-  await httpClientWithResponse(url).then((response) => {
-    if (response.message) {
-      cb('phoneNumber', response.message);
+const JuniorFilter = ({ youthClubs, ...props }) => (
+  <Filter {...props}>
+    <TextInput label="Nimi" source="name" autoFocus />
+    <TextInput label="Nuoren puhelinnumero" source="phoneNumber" autoFocus />
+    <TextInput
+      label="Huoltajan puhelinnumero"
+      source="parentsPhoneNumber"
+      autoFocus
+    />
+    <SelectInput
+      label="Kotinuorisotila"
+      source="homeYouthClub"
+      choices={youthClubs}
+    />
+    <SelectInput label="Tila" source="status" choices={statusChoices} />
+  </Filter>
+);
+
+
+const ResendSMSButton = () => {
+  const record = useRecordContext();
+  const notify = useNotify();
+
+  const resendSMS = async (phoneNumber) => {
+    const url = api.junior.reset;
+    const body = JSON.stringify({
+      phoneNumber,
+    });
+    const options = {
+      method: 'POST',
+      body,
+    };
+    await httpClientWithResponse(url, options).then((response) => {
+      if (response.statusCode < 200 || response.statusCode >= 300) {
+        notify(response.message, 'warning');
+      } else {
+        notify(response.message);
+      }
+    });
+  };
+
+  return null;
+
+  return record.status === 'accepted' || record.status === 'expired' ? (
+    <Button
+      size="small"
+      variant="outlined"
+      onClick={() => resendSMS(record.phoneNumber)}
+    >
+      Lähetä SMS uudestaan
+    </Button>
+  ) : (
+    <Button disabled>Kotisoitto tekemättä</Button>
+  );
+};
+
+const PrintQrCodeButton = () => {
+  const record = useRecordContext();
+
+  const generateQRAndOpen = async (id, owner) => {
+    try {
+      const data = await QRCode.toDataURL(id);
+      const image = new Image();
+      image.src = data;
+      image.width = 400;
+
+      const w = window.open('');
+      setTimeout(() => (w.document.title = `QR-koodi ${owner}`), 0);
+      w.document.write(image.outerHTML);
+      w.document.location = '#';
+      w.document.close();
+    } catch (err) {
+      alert('Virhe QR-koodin luonnissa');
     }
-  });
+  };
+
+  return (
+    <Button
+      size="small"
+      variant="outlined"
+      onClick={() =>
+        generateQRAndOpen(
+          record.id,
+          `${record.firstName} ${record.lastName}`,
+        )
+      }
+    >
+      <>🔍QR</>
+    </Button>
+  );
 };
 
 const DummyPhoneNumberButton = () => {
-  const form = useForm();
+  const { setValue } = useFormContext();
   return (
     <Button
-      variant="contained"
-      color="primary"
+      variant="outline"
       size="small"
-      onClick={() =>
-        getDummyPhoneNumber((field, value) => form.change(field, value))
-      }
+      onClick={() => {
+        httpClientWithResponse(api.junior.dummynumber).then((response) => {
+          if (response.message) {
+            setValue('phoneNumber', response.message);
+          }
+        });
+      }}
     >
       Käytä korvikepuhelinnumeroa
     </Button>
@@ -241,7 +248,7 @@ export const JuniorCreate = (props) => {
           source="phoneNumber"
           validate={required()}
         />
-        <FormDataConsumer>{() => <DummyPhoneNumberButton />}</FormDataConsumer>
+        <DummyPhoneNumberButton />
         {valueOrNull(
           'postCode',
           <TextInput
@@ -301,7 +308,6 @@ export const JuniorCreate = (props) => {
 
 export const JuniorEdit = (props) => {
   const [youthClubs, setYouthClubs] = useState([]);
-  const { isSuperAdmin } = usePermissions();
 
   useEffect(() => {
     const addYouthClubsToState = async () => {
@@ -351,7 +357,7 @@ export const JuniorEdit = (props) => {
           source="phoneNumber"
           validate={required()}
         />
-        <FormDataConsumer>{() => <DummyPhoneNumberButton />}</FormDataConsumer>
+        <DummyPhoneNumberButton />
         {valueOrNull(
           'postCode',
           <TextInput
@@ -390,32 +396,37 @@ export const JuniorEdit = (props) => {
           choices={languages}
           validate={required()}
         />
-        <BooleanInput label="Kuvauslupa" source="photoPermission" className="toggleField"/>
-        <FormDataConsumer>
-          {({ record }) => {
-            const disabled = record.status === 'expired' && !isSuperAdmin;
-            return (
-              <SelectInput
-                disabled={disabled}
-                label="Tila"
-                source="status"
-                choices={statusChoices}
-                validate={required()}
-              />
-            );
-          }}
-        </FormDataConsumer>
-        <FormDataConsumer>
-          {({ formData, record }) =>
-            formData.status === 'accepted' &&
-            (record.status === 'pending' || record.status === 'failedCall') && (
-              <SMSwarning />
-            )
-          }
-        </FormDataConsumer>
+        <BooleanInput label="Kuvauslupa" source="photoPermission" className="toggleField" />
+        <StatusInput />
+        <SMSwarningInput />
       </SimpleForm>
     </Edit>
   );
+};
+
+const StatusInput = () => {
+  const { isSuperAdmin } = usePermissions();
+  const record = useRecordContext();
+  const disabled = record.status === 'expired' && !isSuperAdmin;
+  return (
+    <SelectInput
+      disabled={disabled}
+      label="Tila"
+      source="status"
+      choices={statusChoices}
+      validate={required()}
+    />
+  );
+};
+
+const SMSwarningInput = () => {
+  const record = useRecordContext();
+  const status = useWatch({ name: 'status' });
+
+  if (status === 'accepted' && (record.status === 'pending' || record.status === 'failedCall')) {
+    return <SMSwarning />;
+  }
+  return null;
 };
 
 const languages = [
